requirements.txt
requests
langchain
langchain-community
langchain-core
langchain-ollama
feedparser
wikipedia-api
uvicorn
fastapi
python-dotenv
numpy
matplotlib
scikit-learn
chromadb
PyPDF2
langchain-openai 
sqlalchemy 
psycopg2-binary
crewai
crewai_tools
nltk
PyMuPDF
faiss-cpu 
sentence-transformers
ollama